data-analyzer/requirements.txt
# 數據處理
pandas>=2.1.4
numpy>=1.26.3
scipy>=1.11.4

# 資料庫
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# 機器學習
scikit-learn>=1.3.2
xgboost>=2.0.3
lightgbm>=4.1.0

# 深度學習
torch>=2.1.2
# tensorflow>=2.15.0

# 技術指標
ta-lib-bin>=0.4.28  # 或使用 pandas-ta
pandas-ta>=0.3.14b

# 視覺化
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter
jupyterlab>=4.0.10
ipywidgets>=8.1.1

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.5.3
pyyaml>=6.0.1

# 工具
loguru>=0.7.2
tqdm>=4.66.1

# 報表生成
weasyprint>=60.1  # PDF 生成（可選）
